+++ src/yoob/controller.js
@@ -1,5 +1,5 @@
 /*
- * This file is part of yoob.js version 0.8
+ * This file is part of yoob.js version 0.11-PRE
  * Available from https://github.com/catseye/yoob.js/
  * This file is in the public domain.  See http://unlicense.org/ for details.
  */
@@ -260,12 +260,16 @@ yoob.Controller = function() {
 
     this.performReset = function(state) {
         if (state !== undefined) {
-            this.resetState = state;
+            this.setResetState(state);
         }
         this.reset(this.resetState);
     };
 
     this.reset = function(state) {
         alert("reset() NotImplementedError");
+    };
+
+    this.setResetState = function(state) {
+        this.resetState = state;
     };
 };