+++ HISTORY.markdown
@@ -3,9 +3,9 @@ Changelog
 
 *   version 0.11-PRE
     
-    `yoob.Playfield` and its Views (`yoob.PlayfieldCanvasView`,
-    `yoob.PlayfieldHTMLView`) take configuration dictionaries as their
-    single argument to `init()`.
+    `yoob.Playfield`, its Views (`yoob.PlayfieldCanvasView`,
+    `yoob.PlayfieldHTMLView`), and `yoob.Cursor` all take configuration
+    dictionaries as their single argument to `init()`.
     
     `yoob.Cursors` are now attached to `yoob.Playfield` objects, and no
     longer attached to their Views.